add sensor ccs811, use TickerSceduler

+++ include/max7219display.h
@@ -3,5 +3,5 @@
 
 
 void max7219displayInit (void);
-void printTwoNumb(int, int);
+void printTwoNumb(float, int);
 void printSymbolTermo(void);
